app/requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
#tensorflow==2.1
keras==2.3.1
opencv-python
h5py
imgaug
IPython[all]

#pandas==1.0.1
#numpy==1.18.1
#schema==0.6.8
#requests==2.22.0
#Flask==0.12.2
#concurrent_log_handler==0.9.16
#scikit_learn==0.22.0
#gunicorn==19.9.0


#click==6.7
#concurrent-log-handler==0.9.9
Flask==1.0.2
Flask-Testing==0.7.1
itsdangerous==0.24
#Jinja2==2.10
MarkupSafe==1.0
#numpy==1.14.5